[FIXED JENKINS-23228]
Documented this feature and integrated a newer version of Stapler.

I took the opportunity to reorder the content in the section, so that
the tree parameter becomes the main subject.

At this point, the depth parameter should be only used in a limited
circumstance, so it should be treated as such.

Conflicts:
	changelog.html
	core/pom.xml